decomp: reloc deep-dive fixes — previously-blocked/unresolved functions (17)
Whole-function reconstruction + compiler-validated behavior-preserving edits
for functions that resisted the reloc-masked burn-down (waves 4-10). Each was
validated with tools/revcomp/validate_reloc_fix.py (isolate build, name_only)
and confirmed in a full in-tree build; overall Milo Engine fuzzy holds and
several functions jump materially. base->new (name_only):

- ??0DxMesh@@IAA@XZ (Mesh.cpp): merge three if-scoped function-local static
  D3DVERTEXELEMENT9 arrays into one pooled static (single base register +
  offsets), &sVertexElements[0]/[8]/[15]. 93.85 -> 99.64.
- ?InitInfo@StandardStream (StandardStream.cpp): correct wrong division
  (unkec = (float)i4/sampleRate), drop address-forcing reference aliases,
  hoist mChanParams[i]->mSlipEnabled. 82.09 -> 93.72.
- ?CalculateFaderVolume@ThreeDSound (ThreeDSound.cpp): shape dispatch
  if/else-chain -> switch, radius guard split out. 84.24 -> 88.91.
- ?SetPersistentPool@RndParticleSys (Part.cpp): null-terminate the LAST list
  element (reuse its address) instead of one-past-the-end; corrects a latent
  OOB write. 93.85 -> 95.78.
- ?Load@DancerSequence (DancerSequence.cpp): legacy rev<3 path reads a
  Key<float> via BinStream& (d.stream), not a std::vector<float>; fixes a
  stream-desync bug for old-format files. 90.20 -> 93.59.
- ??0RndPartLauncher (PartLauncher.cpp, Part.h, Part.cpp): default-construct
  mPartOverride (was copy-from-global) + nothrow ctor for clean call-site EH
  codegen. 87.22 -> 90.39.
- ?FitTextEllipsis@RndText (Text.cpp): inline wide-string literals L"..." /
  L" .," (pooled anon COMDAT, corrects break-char set), lines-before-wideChars
  decl order, charWidths sized numChars+5. 88.13 -> 88.49.
- ?Process@WahEffect (WahEffect.cpp): delay the mSweepRate read to just before
  its compare so the freq-multiply constant-pool order matches. 91.55 -> 92.62.
- ?Process@DistortionEffect (DistortionEffect.cpp): right-channel pointer
  advances by the 2-float stereo stride (right += 2). 97.31 -> 98.21.
- ?TakeShot@HiResScreen (HiResScreen.cpp): remove (int) casts on unsigned
  members so the tile-count compare is unsigned (cmplw). 88.01 -> 89.22.
- ?createOrReuse@Voice (Voice.cpp): early-return guard with a flat body.
  94.42 -> 94.45.
- ?SetReleaseSmoothing@Synapse@DSP (Synapse_dsp.cpp): call the ICF-fold
  retained target (ExposureRecipe::SetMinIntegrationTime, same field/offset).
  99.50 -> 99.63.
- ?fft_matrix_forward_columnwise (FFT.cpp): pooled __vmx@ vector literals with
  the correct sign vector {1,-1,1,-1}. 54.70 -> 54.76.
- ?Create@RndBitmap (Bitmap.h): #line override so the inlined debug-allocator
  __FILE__ uses the native all-backslash path. 99.72 -> 99.84.
- ?SavePersistentObjects@WorldInstance (Instance.cpp, Dir.h): restore the
  ObjDirItr::Ptr() accessor the MILO_ASSERT stringizer expected. 98.31 -> 98.38.
- ?Execute@FlowSetProperty (FlowSetProperty.h): #line 1 basename so inline
  __FILE__ (POOL_OVERLOAD/OBJ_MEM_OVERLOAD) matches the original .cpp string.
  97.36 -> 97.38.
- ?StartPlayback@MicXbox (synth_xbox/Voice.h): #line override restoring the
  original synth360 all-backslash path. 96.97 -> 97.07.

All edits are behavior-preserving (or behavior-correcting: DancerSequence
stream-desync, SetPersistentPool OOB). Residuals per function are certified
floors (regalloc/scheduling, /OPT:ICF folds, anon-namespace hash, constant-pool
layout, __FILE__ build-env) recorded in
decomp-synth docs/plans/reverse-compilation/SESSION_2026-07-04_RELOC_DEEPDIVE.md.

+++ src/system/rndobj/HiResScreen.cpp
@@ -162,7 +162,7 @@ void HiResScreen::TakeShot(const char *c, int i) {
     } else {
         mAccumWidth = i * (TheRnd.Width() - 480);
         mAccumHeight = i * TheRnd.Height() - i * 270;
-        if ((int)mAccumWidth < TheRnd.Width() || (int)mAccumHeight < TheRnd.Height()) {
+        if (mAccumWidth < TheRnd.Width() || mAccumHeight < TheRnd.Height()) {
             MILO_NOTIFY(MakeString("HiResScreenshot requires more tiles (%d specified)", i));
             mActive = false;
         } else {
